+++ versions/v21/projects/sdcorejs-angular/components/org-chart/index.ts
@@ -1,0 +1,3 @@
+export * from './src/org-chart.component';
+export * from './src/org-chart-item-def.directive';
+export * from './src/org-chart.model';